Bundle geth & swarm binaries in electron

+++ app/index.js
@@ -1,13 +1,15 @@
-const { app, BrowserWindow, Menu, ipcMain } = require('electron')
+const { app, BrowserWindow, dialog, ipcMain, Menu } = require('electron')
 const { appReady, is } = require('electron-util')
 const Store = require('electron-store')
 const getPort = require('get-port')
 const createOnyxServer = require('onyx-server').default
-const path = require('path')
 const StaticServer = require('static-server')
 const url = require('url')
+const path = require('path')
 
 const { config } = require(path.join(__dirname, 'package.json'))
+const swarm = require('./swarm')
+
 const SWARM_WS_URL =
   process.env.SWARM_WS_URL ||
   (config && config.swarmWsUrl) ||
@@ -17,20 +19,52 @@ const SWARM_HTTP_URL =
   (config && config.swarmHttpUrl) ||
   'http://localhost:8500'
 
+let appServer, loadingWindow, mainWindow
+
+const shouldQuit = app.makeSingleInstance(() => {
+  if (mainWindow) {
+    if (mainWindow.isMinimized()) mainWindow.restore()
+    mainWindow.focus()
+  }
+})
+
+if (shouldQuit) {
+  app.quit()
+}
+
+const store = new Store({ name: is.development ? 'onyx-dev' : 'onyx' })
+
+const clearEventListeners = () => {
+  if (mainWindow) {
+    mainWindow.removeListener('ready-to-show', showFunc)
+    mainWindow.removeListener('closed', closedFunc)
+  }
+}
+
 const menu = Menu.buildFromTemplate([
   {
     label: is.macos ? 'Onyx' : 'File',
     submenu: [
       {
         label: 'Reset',
-        click: () => {
-          store.delete('wsUrl')
-          if (mainWindow != null) {
-            clearEventListeners()
-            mainWindow.close()
-            mainWindow = null
-          }
-          start()
+        click: (item, focusedWindow) => {
+          dialog.showMessageBox(
+            focusedWindow,
+            {
+              type: 'warning',
+              message:
+                'Resetting the application will clear ALL data and your local Swarm identity. Are you sure you want to continue?',
+              cancelId: 0,
+              buttons: ['Cancel', 'Confirm'],
+            },
+            async index => {
+              if (index === 1) {
+                store.delete('wsUrl')
+                await swarm.reset()
+                start()
+              }
+            },
+          )
         },
       },
       {
@@ -84,10 +118,6 @@ const menu = Menu.buildFromTemplate([
   },
 ])
 
-let appServer, loadingWindow, mainWindow, connectionError
-
-const store = new Store({ name: is.development ? 'onyx-dev' : 'onyx' })
-
 const createMainWindow = async url => {
   await appReady
 
@@ -96,12 +126,13 @@ const createMainWindow = async url => {
   mainWindow = new BrowserWindow({ width: 800, height: 600, show: false })
 
   mainWindow.loadURL(url)
-  
+
   showFunc = () => {
     mainWindow.show()
   }
-  
+
   closedFunc = () => {
+    swarm.stop()
     mainWindow = null
   }
 
@@ -152,16 +183,41 @@ const start = async () => {
   const storedWsUrl = store.get('wsUrl')
   if (storedWsUrl) {
     if (storedWsUrl === 'local') {
-      // Setup a local Graphql server
+      let errorMsg
+
       try {
-        const serverPort = await startLocalOnyxServer(appPort)
-        const wsUrl = `ws://localhost:${serverPort}/graphql`
-        const httpUrl = `http://localhost:${serverPort}`
-        appUrl = appUrl + `/?wsUrl=${wsUrl}&httpUrl=${httpUrl}`
-      } catch (err) {
-        console.warn('err: ', err)
-        const errorMsg = 'There was an issue starting local GraphQL server, you may want to check you have a swarm node running on default port 8546, or that you specified the correct port if not using default'
-        appUrl = appUrl + `/?wsUrl=${storedWsUrl}&connectionError=${errorMsg}`
+        await swarm.setup()
+      } catch (e) {
+        console.log(e.stack)
+        errorMsg =
+          'There was an issue setting up the Geth account, you may want to check if your configuration is valid'
+      }
+
+      if (!errorMsg) {
+        try {
+          await swarm.start()
+        } catch (e) {
+          console.log(e.stack)
+          errorMsg =
+            'There was an issue starting local Swarm node, you may want to check if you have a Swarm node running already, or if the port 30399 is open'
+        }
+      }
+
+      if (!errorMsg) {
+        try {
+          const serverPort = await startLocalOnyxServer(appPort)
+          const wsUrl = `ws://localhost:${serverPort}/graphql`
+          const httpUrl = `http://localhost:${serverPort}`
+          appUrl += `/?wsUrl=${wsUrl}&httpUrl=${httpUrl}`
+        } catch (e) {
+          console.log(e.stack)
+          errorMsg =
+            'There was an issue starting local GraphQL server, you may want to check you have a Swarm node running on default port 8546, or that you specified the correct port if not using default'
+        }
+      }
+
+      if (errorMsg) {
+        appUrl += `/?wsUrl=${storedWsUrl}&connectionError=${errorMsg}`
         if (appServer != null) {
           appServer.stop()
         }
@@ -169,18 +225,14 @@ const start = async () => {
     } else {
       // Use stored remote server url
       let domain
-      if (storedWsUrl.indexOf('://') > -1) {
+      if (storedWsUrl.indexOf('://') !== -1) {
         domain = storedWsUrl.split('/')[2]
       } else if (storedWsUrl.indexOf('/') !== -1) {
         domain = storedWsUrl.split('/')[0]
       }
-      if (!domain) {
-        const errorMsg = 'Invalid ws url'
-        appUrl = appUrl + `/?wsUrl=${storedWsUrl}&connectionError=${errorMsg}`
-      } else {
-        const httpUrl = `http://${domain}`
-        appUrl = appUrl + `/?wsUrl=${storedWsUrl}&httpUrl=${httpUrl}`
-      }
+      appUrl += domain
+        ? `/?wsUrl=${storedWsUrl}&httpUrl=http://${domain}`
+        : `/?wsUrl=${storedWsUrl}&connectionError=Invalid ws url`
     }
   }
 
@@ -195,17 +247,13 @@ const start = async () => {
       createMainWindow(appUrl)
     }
   })
-  
-  app.on('certificate-error', (event, webContents, url, error, certificate, callback) => {
-    console.warn('cert error: ', error)
-  })
-}
 
-const clearEventListeners = () => {
-  if (mainWindow) {
-    mainWindow.removeListener('ready-to-show', showFunc)
-    mainWindow.removeListener('closed', closedFunc)
-  }
+  app.on(
+    'certificate-error',
+    (event, webContents, url, error, certificate, callback) => {
+      console.warn('cert error: ', error)
+    },
+  )
 }
 
 ipcMain.on('onSetWsUrl', (e, url) => {
